EVEREST-610 Deprecated install operators command
